chore: remove unused tests from `llm_chat/tests.py` to clean up codebase and improve maintainability
