+++ llm_chat/domain/service/completion/riddle.py
@@ -11,7 +11,7 @@ from llm_chat.domain.valueobject.completion.riddle import (
     Gender,
     RiddleResponse,
     RiddleEvaluation,
-    RiddleItem,
+    Riddle,
 )
 from llm_chat.domain.valueobject.completion.use_case import UseCaseType
 
@@ -36,7 +36,7 @@ class RiddleChatService(BaseLLMTask):
     @staticmethod
     def get_prompt(
         gender: Gender,
-        riddle_set: list[RiddleItem],
+        riddle_set: list[Riddle],
         current_index: int | None = None,
     ) -> str:
         riddle_count = len(riddle_set)
@@ -86,7 +86,7 @@ class RiddleChatService(BaseLLMTask):
     def create_initial_prompt(
         user_message: MessageDTO,
         gender: Gender,
-        riddle_set: list[RiddleItem],
+        riddle_set: list[Riddle],
     ) -> list[MessageDTO]:
         """
         初期プロンプト（システムメッセージと初回のユーザーメッセージ）を生成します。
@@ -114,7 +114,7 @@ class RiddleChatService(BaseLLMTask):
         # システムメッセージを先頭に含めて返す
         return [system_message, first_user_message]
 
-    def execute(self, login_user: User, riddle_set: list[RiddleItem]) -> RiddleResponse:
+    def execute(self, login_user: User, riddle_set: list[Riddle]) -> RiddleResponse:
         """
         評価プロンプトを送信し、結果を RiddleResponse に変換します。
         """
@@ -133,11 +133,17 @@ class RiddleChatService(BaseLLMTask):
 
 ### 評価の指示
 - 挨拶、説明、前置きなどは一切不要です。
-- JSON配列（[...]）のみを返してください。
-- フォーマットは必ず以下の判定結果例に従ってください。
+- JSONオブジェクト（{{...}}）のみを返してください。
+- フォーマットは必ず以下のJSON構造に従ってください。
 
-判定結果例:
-[{{"viewpoint": "論理的思考力", "score": 80, "judge": "合格"}}, {{"viewpoint": "洞察力", "score": 40, "judge": "不合格"}}]
+JSON構造:
+{{
+  "correctness": 3,
+  "reasoning": 4,
+  "creativity": 2,
+  "rebuttal": 0,
+  "comment": "コメント"
+}}
 """.strip(),
         )
 
@@ -162,29 +168,22 @@ class RiddleChatService(BaseLLMTask):
                 cleaned_content = "\n".join(lines).strip()
 
             eval_data = json.loads(cleaned_content)
-            # eval_data はリスト形式 [{...}, {...}] と想定
-            if not isinstance(eval_data, list):
-                raise ValueError("LLM returned non-list format for evaluations")
+            # eval_data は辞書形式 {...} と想定
+            if not isinstance(eval_data, dict):
+                raise ValueError("LLM returned non-dict format for evaluation")
 
-            evaluations = []
-            for item in eval_data:
-                if isinstance(item, dict):
-                    evaluations.append(RiddleEvaluation(**item))
-                else:
-                    # item が辞書でない場合はスキップするか、ValueErrorを投げてパース失敗扱いにする
-                    raise ValueError(f"Evaluation item is not a dict: {type(item)}")
+            evaluation = RiddleEvaluation(**eval_data)
 
             return RiddleResponse(
                 answer=raw_content,
-                evaluations=evaluations,
+                evaluation=evaluation,
                 explanation="なぞなぞの評価結果です。",
             )
         except (json.JSONDecodeError, ValueError) as e:
             # パース失敗時のフォールバックまたはエラーハンドリング
-            # ここで「残心」を持ってエラーを処理
             return RiddleResponse(
                 answer=raw_content,
-                evaluations=[],
+                evaluation=None,
                 explanation=f"評価結果のパースに失敗しました: {str(e)}",
             )
 
@@ -195,6 +194,6 @@ class RiddleChatService(BaseLLMTask):
         パース失敗時は explanation を含めます。
         """
         text = response.to_bullet_points()
-        if not response.evaluations:
+        if not response.evaluation:
             text += f"\n- {response.explanation}"
         return text
