--- llm_chat/tests.py
@@ -1,1026 +1,0 @@
-import time
-from datetime import timedelta
-from django.test import TestCase, RequestFactory
-from django.contrib.sessions.backends.db import SessionStore
-from django.contrib.auth.models import User
-from lib.llm.valueobject.completion import RoleType
-from lib.llm.valueobject.config import OpenAIGptConfig, ModelName
-from django.core.files.uploadedfile import SimpleUploadedFile
-from django.utils import timezone
-from llm_chat.models import ChatLogs, RiddleQuestion
-from llm_chat.views import IndexView
-from llm_chat.domain.valueobject.completion.chat import MessageDTO
-from llm_chat.domain.valueobject.completion.riddle import Gender, GenderType
-from llm_chat.domain.valueobject.completion.use_case import UseCaseType
-from llm_chat.domain.repository.completion.chat import ChatLogRepository
-from llm_chat.domain.service.completion.chat import ChatService
-from llm_chat.domain.service.completion.riddle import RiddleChatService
-from llm_chat.domain.use_case.completion.chat import LlmChatUseCase
-from llm_chat.domain.use_case.completion.multimedia import (
-    OpenAIDalleUseCase,
-    OpenAITextToSpeechUseCase,
-    OpenAISpeechToTextUseCase,
-)
-from llm_chat.domain.use_case.completion.riddle import RiddleUseCase
-from unittest.mock import patch, MagicMock
-
-
-class ChatModelAndRepositoryTest(TestCase):
-    def setUp(self):
-        self.user = User.objects.create_user(username="test_user")
-
-    def test_chat_logs_to_message_dto(self):
-        """
-        [シナリオ]
-        1. ChatLogs エンティティを作成 (role='user', content='Hello', use_case_type=UseCaseType.OPENAI_GPT)
-        2. to_message_dto() を呼び出して MessageDTO に変換
-        3. 期待値: 各フィールドが正しくマッピングされ、use_case_type が UseCaseType.OPENAI_GPT であること
-        """
-        log = ChatLogs.objects.create(
-            user=self.user,
-            role=RoleType.USER.value,
-            content="Hello",
-            model_name=ModelName.GPT_4O,
-        )
-        dto = log.to_message_dto()
-        self.assertEqual(dto.content, "Hello")
-        self.assertEqual(dto.model_name, ModelName.GPT_4O)
-        self.assertEqual(dto.use_case_type, UseCaseType.OPENAI_GPT)
-
-    def test_repository_insert_and_find(self):
-        """
-        [シナリオ]
-        1. ユーザーメッセージを模した MessageDTO を作成 (role='assistant', use_case_type=UseCaseType.RIDDLE)
-        2. ChatLogRepository.insert() を使用して DB に保存
-        3. find_chat_history() でそのユーザーの履歴を取得
-        4. 期待値: 取得した履歴が1件であり、内容と use_case_type が一致すること
-        """
-        dto = MessageDTO(
-            user=self.user,
-            role=RoleType.ASSISTANT,
-            content="AI response",
-            model_name=ModelName.GPT_4O,
-            use_case_type=UseCaseType.RIDDLE,
-        )
-        ChatLogRepository.insert(dto)
-
-        history = ChatLogRepository.find_chat_history(self.user)
-        self.assertEqual(len(history), 1)
-        self.assertEqual(history[0].content, "AI response")
-        self.assertEqual(history[0].use_case_type, UseCaseType.RIDDLE)
-
-
-class ChatLogicTest(TestCase):
-    def setUp(self):
-        self.user = User.objects.create_user(username="logic_user")
-        # デフォルトの問題を登録
-        RiddleQuestion.objects.create(
-            question_text="はじめは4本足、途中から2本足、最後は3本足。それは何でしょう？",
-            answer_text="人間",
-            order=1,
-        )
-        RiddleQuestion.objects.create(
-            question_text="私は黒い服を着て、赤い手袋を持っている。夜には立っているが、朝になると寝る。何でしょう？",
-            answer_text="たいまつ",
-            order=2,
-        )
-
-    def test_get_chat_history_normal(self):
-        """
-        [シナリオ: 通常チャット]
-        1. 過去の履歴がない状態で、ユーザーメッセージ (use_case_type=UseCaseType.OPENAI_GPT) を受け取る
-        2. get_chat_history() を実行
-        3. 期待値: 履歴リストにユーザーメッセージのみが含まれ、use_case_type が UseCaseType.OPENAI_GPT であること
-        """
-        user_message = MessageDTO(
-            user=self.user,
-            role=RoleType.USER,
-            content="Normal message",
-            model_name=ModelName.GPT_4O,
-            use_case_type=UseCaseType.OPENAI_GPT,
-        )
-        history = ChatService.get_chat_history(
-            user_message, use_case_type=UseCaseType.OPENAI_GPT
-        )
-        # 履歴が空なので、user_messageのみが保存される
-        self.assertEqual(len(history), 1)
-        self.assertEqual(history[0].content, "Normal message")
-        self.assertEqual(history[0].use_case_type, UseCaseType.OPENAI_GPT)
-
-    def test_get_chat_history_riddle_first(self):
-        """
-        [シナリオ: なぞなぞ初回開始]
-        1. 過去の履歴がない状態で、ユーザーメッセージ (use_case_type=UseCaseType.RIDDLE) を受け取る
-        2. get_chat_history(use_case_type=UseCaseType.RIDDLE) を実行
-        3. 期待値:
-           - 内部的にシステムメッセージが生成され、履歴リストの先頭に追加されること (計2通)
-           - ユーザーメッセージのみが DB に保存されること
-        """
-        user_message = MessageDTO(
-            user=self.user,
-            role=RoleType.USER,
-            content="なぞなぞスタート",
-            model_name=ModelName.GPT_5_MINI,
-            use_case_type=UseCaseType.RIDDLE,
-        )
-        # 初回：システムメッセージ（非保存）と初回ユーザーメッセージ（保存）
-        history = ChatService.get_chat_history(
-            user_message,
-            use_case_type=UseCaseType.RIDDLE,
-            gender=Gender(GenderType.MAN),
-        )
-        self.assertEqual(len(history), 2)
-        self.assertEqual(history[0].role, RoleType.SYSTEM)
-        self.assertEqual(history[1].role, RoleType.USER)
-        self.assertEqual(history[1].use_case_type, UseCaseType.RIDDLE)
-
-        # DBには初回ユーザーメッセージが保存されているはず
-        db_logs = ChatLogs.objects.filter(user=self.user)
-        self.assertEqual(db_logs.count(), 1)
-        self.assertEqual(db_logs[0].role, RoleType.USER.value)
-
-    @patch("lib.llm.service.completion.LlmCompletionService.retrieve_answer")
-    def test_riddle_use_case_extra_question_removal(self, mock_retrieve):
-        """
-        [シナリオ: なぞなぞ終了時の余計な質問除去]
-        1. LLMが規定回数終了時に「第3問」を出題しようとするケースを模倣
-        2. RiddleUseCase.execute() を実行
-        3. 期待値:
-           - 「第3問」以降が除去され、終了定型文で終わっていること
-        """
-        # ユーザーの発言を蓄積（3回目の発言＝2問目への回答を想定）
-        ChatLogRepository.insert(
-            MessageDTO(
-                user=self.user,
-                role=RoleType.USER,
-                content="スタート",
-                use_case_type=UseCaseType.RIDDLE,
-            )
-        )
-        ChatLogRepository.insert(
-            MessageDTO(
-                user=self.user,
-                role=RoleType.ASSISTANT,
-                content="第1問...",
-                use_case_type=UseCaseType.RIDDLE,
-            )
-        )
-        ChatLogRepository.insert(
-            MessageDTO(
-                user=self.user,
-                role=RoleType.USER,
-                content="答え1",
-                use_case_type=UseCaseType.RIDDLE,
-            )
-        )
-        ChatLogRepository.insert(
-            MessageDTO(
-                user=self.user,
-                role=RoleType.ASSISTANT,
-                content="第2問...",
-                use_case_type=UseCaseType.RIDDLE,
-            )
-        )
-
-        # LLMの回答を模倣（余計な第3問を含む）
-        assistant_content = (
-            "正解です！たいまつで合っています。\n"
-            "では第3問です。\n"
-            "僕は呼吸をするけど生きていない...\n"
-            f"ご回答をどうぞ。\n\n"
-            f"{RiddleChatService.RIDDLE_END_MESSAGE}"
-        )
-        # 1回目の呼び出し（回答生成）と2回目の呼び出し（評価生成）
-        mock_retrieve.side_effect = [
-            MagicMock(answer=assistant_content),
-            MagicMock(
-                answer='[{"viewpoint": "テスト", "score": 100, "judge": "合格"}]'
-            ),
-        ]
-
-        config = OpenAIGptConfig(
-            api_key="fake", max_tokens=100, model=ModelName.GPT_5_MINI
-        )
-        use_case = RiddleUseCase(config)
-
-        result = use_case.execute(self.user, "答え2", gender=Gender(GenderType.MAN))
-
-        # 「第3問」や「僕は呼吸をするけど...」が消えていることを確認
-        self.assertIn("正解です！たいまつで合っています。", result.content)
-        self.assertIn(RiddleChatService.RIDDLE_END_MESSAGE, result.content)
-        self.assertNotIn("第3問", result.content)
-        self.assertNotIn("僕は呼吸をするけど", result.content)
-        self.assertIn("テスト: 100点 (合格)", result.content)
-
-    @patch("lib.llm.service.completion.LlmCompletionService.retrieve_answer")
-    def test_riddle_use_case_end_detection_invalid_json(self, mock_retrieve):
-        """
-        [シナリオ: なぞなぞ終了判定（不正なJSON）]
-        1. LLM の回答がリスト形式だが中身が文字列であるケースを模倣
-        2. RiddleUseCase.execute() を実行
-        3. 期待値:
-           - TypeError を起こさず、空の評価結果で終了すること
-        """
-        # 不正なJSON（リストの中に文字列）を模倣
-        mock_retrieve.return_value = MagicMock(answer='["invalid", "string", "items"]')
-
-        config = OpenAIGptConfig(
-            api_key="fake", max_tokens=100, model=ModelName.GPT_5_MINI
-        )
-        use_case = RiddleUseCase(config)
-
-        # 3回目のユーザー発言相当のコンテキストを作る
-        ChatLogRepository.insert(
-            MessageDTO(
-                user=self.user,
-                role=RoleType.USER,
-                content="スタート",
-                use_case_type=UseCaseType.RIDDLE,
-            )
-        )
-        ChatLogRepository.insert(
-            MessageDTO(
-                user=self.user,
-                role=RoleType.ASSISTANT,
-                content="質問1...",
-                use_case_type=UseCaseType.RIDDLE,
-            )
-        )
-        ChatLogRepository.insert(
-            MessageDTO(
-                user=self.user,
-                role=RoleType.USER,
-                content="答え1",
-                use_case_type=UseCaseType.RIDDLE,
-            )
-        )
-        ChatLogRepository.insert(
-            MessageDTO(
-                user=self.user,
-                role=RoleType.ASSISTANT,
-                content="質問2...",
-                use_case_type=UseCaseType.RIDDLE,
-            )
-        )
-
-        # evaluate 内部での LLM 実行（評価 JSON 取得）がこの mock_retrieve を使う
-        result = use_case.execute(
-            self.user, "答えはたいまつです", gender=Gender(GenderType.MAN)
-        )
-
-        self.assertIn(RiddleChatService.RIDDLE_END_MESSAGE, result.content)
-        self.assertIn("評価結果のパースに失敗しました", result.content)
-        self.assertEqual(result.use_case_type, UseCaseType.RIDDLE)
-
-    @patch("lib.llm.service.completion.LlmCompletionService.retrieve_answer")
-    def test_riddle_use_case_end_detection(self, mock_retrieve):
-        """
-        [シナリオ: なぞなぞ終了判定]
-        1. LLM の回答に終了キーワード (RIDDLE_END_MESSAGE) が含まれるケースを模倣
-        2. RiddleUseCase.execute() を実行
-        3. 期待値:
-           - 回答内容に終了メッセージが含まれていること
-           - 回答後の評価結果が含まれていること
-           - メッセージの use_case_type が UseCaseType.RIDDLE であること
-        """
-        # 終了メッセージを含む回答を模倣
-        mock_retrieve.return_value = MagicMock(
-            answer=f"正解です！ {RiddleChatService.RIDDLE_END_MESSAGE}"
-        )
-
-        config = OpenAIGptConfig(
-            api_key="fake", max_tokens=100, model=ModelName.GPT_5_MINI
-        )
-        use_case = RiddleUseCase(config)
-
-        # RiddleUseCase.execute は内部で evaluate を呼ぶ（さらに LLM 実行）
-        # 簡単のため、evaluate もモック化するか、retrieve_answer を 2回返すように設定
-        with patch(
-            "llm_chat.domain.service.completion.chat.ChatService.evaluate"
-        ) as mock_eval:
-            mock_eval.return_value = "\n【評価結果】\n- 論理的思考力: 100点 (合格)"
-
-            result = use_case.execute(
-                self.user, "答えは人間です", gender=Gender(GenderType.MAN)
-            )
-
-            self.assertIn(RiddleChatService.RIDDLE_END_MESSAGE, result.content)
-            self.assertIn("評価結果", result.content)
-            self.assertEqual(result.use_case_type, UseCaseType.RIDDLE)
-
-    @patch("lib.llm.service.completion.LlmCompletionService.retrieve_answer")
-    def test_riddle_use_case_forced_end(self, mock_retrieve):
-        """
-        [シナリオ: なぞなぞ強制終了]
-        1. LLM の回答に終了キーワードが含まれないが、3回目の発言（質問2への回答）である場合
-        2. RiddleUseCase.execute() を実行
-        3. 期待値:
-           - 回答内容に終了メッセージが強制的に追加されていること
-           - 評価結果が含まれていること
-        """
-        # 終了メッセージを含まない回答
-        mock_retrieve.return_value = MagicMock(
-            answer="それはたいまつですね。正解です！"
-        )
-
-        # 過去に2回ユーザー発言がある状態を作る（今回で3回目）
-        ChatLogRepository.insert(
-            MessageDTO(
-                user=self.user,
-                role=RoleType.USER,
-                content="スタート",
-                use_case_type=UseCaseType.RIDDLE,
-            )
-        )
-        ChatLogRepository.insert(
-            MessageDTO(
-                user=self.user,
-                role=RoleType.ASSISTANT,
-                content="質問1...",
-                use_case_type=UseCaseType.RIDDLE,
-            )
-        )
-        ChatLogRepository.insert(
-            MessageDTO(
-                user=self.user,
-                role=RoleType.USER,
-                content="答え1",
-                use_case_type=UseCaseType.RIDDLE,
-            )
-        )
-        ChatLogRepository.insert(
-            MessageDTO(
-                user=self.user,
-                role=RoleType.ASSISTANT,
-                content="質問2...",
-                use_case_type=UseCaseType.RIDDLE,
-            )
-        )
-
-        config = OpenAIGptConfig(
-            api_key="fake", max_tokens=100, model=ModelName.GPT_5_MINI
-        )
-        use_case = RiddleUseCase(config)
-
-        with patch(
-            "llm_chat.domain.service.completion.chat.ChatService.evaluate"
-        ) as mock_eval:
-            mock_eval.return_value = "\n【評価結果】\n- 評価1: 100点"
-
-            # 3回目のユーザー発言
-            result = use_case.execute(
-                self.user, "それはたいまつです", gender=Gender(GenderType.MAN)
-            )
-
-            self.assertIn(RiddleChatService.RIDDLE_END_MESSAGE, result.content)
-            self.assertIn("評価結果", result.content)
-            self.assertEqual(result.use_case_type, UseCaseType.RIDDLE)
-
-    @patch("lib.llm.service.completion.LlmCompletionService.retrieve_answer")
-    def test_riddle_use_case_custom_max_turns(self, mock_retrieve):
-        """
-        [シナリオ: なぞなぞカスタム最大回数]
-        1. max_turns=2 (1問で終了) として設定
-        2. 2回目のユーザー発言（質問1への回答）を行う
-        3. 期待値: 2回目で強制終了し、評価結果が含まれること
-        """
-        mock_retrieve.return_value = MagicMock(answer="それは人間ですね。正解です！")
-
-        # 1回目のユーザー発言がある状態（スタート）
-        ChatLogRepository.insert(
-            MessageDTO(
-                user=self.user,
-                role=RoleType.USER,
-                content="スタート",
-                use_case_type=UseCaseType.RIDDLE,
-            )
-        )
-        ChatLogRepository.insert(
-            MessageDTO(
-                user=self.user,
-                role=RoleType.ASSISTANT,
-                content="質問1...",
-                use_case_type=UseCaseType.RIDDLE,
-            )
-        )
-
-        config = OpenAIGptConfig(
-            api_key="fake", max_tokens=100, model=ModelName.GPT_5_MINI
-        )
-        # max_turns=2 に設定
-        use_case = RiddleUseCase(config, max_turns=2)
-
-        with patch(
-            "llm_chat.domain.service.completion.chat.ChatService.evaluate"
-        ) as mock_eval:
-            mock_eval.return_value = "\n【評価結果】\n- 評価1: 100点"
-
-            # 2回目のユーザー発言
-            result = use_case.execute(
-                self.user, "それは人間です", gender=Gender(GenderType.MAN)
-            )
-
-            self.assertIn(RiddleChatService.RIDDLE_END_MESSAGE, result.content)
-            self.assertIn("評価結果", result.content)
-            # プロンプトに "1つの問題を" が含まれているか確認（内部的になぞなぞの数が 1 になっているか）
-            # ただし、retrieve_answer に渡されたメッセージの中身を確認する必要がある
-
-    @patch("lib.llm.service.completion.LlmCompletionService.retrieve_answer")
-    def test_llm_chat_use_case_normal(self, mock_retrieve):
-        """
-        [シナリオ: 通常チャットユースケース]
-        1. LlmChatUseCase を使用してユーザーメッセージを送信
-        2. 期待値:
-           - LLM の回答内容が正しく取得されること
-           - 使用モデル名が設定値と一致すること
-           - use_case_type が UseCaseType.OPENAI_GPT であること
-           - ユーザーとアシスタントの計2通が DB に保存されること
-        """
-        mock_retrieve.return_value = MagicMock(answer="AIの回答です")
-        config = OpenAIGptConfig(
-            api_key="fake", max_tokens=100, model=ModelName.GPT_5_MINI
-        )
-        use_case = LlmChatUseCase(config)
-        result = use_case.execute(self.user, "こんにちは")
-
-        self.assertEqual(result.content, "AIの回答です")
-        self.assertEqual(result.model_name, ModelName.GPT_5_MINI)
-        self.assertEqual(result.use_case_type, UseCaseType.OPENAI_GPT)
-        self.assertEqual(
-            ChatLogs.objects.filter(user=self.user).count(), 2
-        )  # User + Assistant
-
-    @patch("lib.llm.service.completion.LlmCompletionService.retrieve_answer")
-    def test_riddle_use_case_normal(self, mock_retrieve):
-        """
-        [シナリオ: なぞなぞユースケース]
-        1. RiddleUseCase を使用してなぞなぞを開始
-        2. 期待値:
-           - 回答内容が取得され、use_case_type が UseCaseType.RIDDLE であること
-           - ユーザーとアシスタントの計2通が DB に保存されること (システムメッセージは保存されない)
-        """
-        mock_retrieve.return_value = MagicMock(answer="それは人間ですか？")
-        config = OpenAIGptConfig(
-            api_key="fake", max_tokens=100, model=ModelName.GPT_5_MINI
-        )
-        use_case = RiddleUseCase(config)
-        result = use_case.execute(self.user, "スタート", gender=Gender(GenderType.MAN))
-
-        self.assertEqual(result.content, "それは人間ですか？")
-        self.assertEqual(result.use_case_type, UseCaseType.RIDDLE)
-        # 初回なぞなぞ：System(非保存), User, Assistant の計2通がDBへ
-        self.assertEqual(ChatLogs.objects.filter(user=self.user).count(), 2)
-
-    @patch("lib.llm.service.completion.LlmCompletionService.retrieve_answer")
-    def test_riddle_use_case_no_questions_raises_error(self, mock_retrieve):
-        """
-        [シナリオ: 問題が未登録の場合の挙動]
-        1. RiddleQuestion をすべて削除
-        2. RiddleUseCase.execute() を実行
-        3. 期待値: ValueError が発生し、適切なメッセージが返されること
-        """
-        RiddleQuestion.objects.all().delete()
-        config = OpenAIGptConfig(
-            api_key="fake", max_tokens=100, model=ModelName.GPT_5_MINI
-        )
-        use_case = RiddleUseCase(config)
-
-        with self.assertRaisesRegex(
-            ValueError,
-            "なぞなぞの問題が登録されていません。管理画面から問題を登録してください。",
-        ):
-            use_case.execute(self.user, "スタート", gender=Gender(GenderType.MAN))
-
-    def test_get_chat_history_riddle_no_gender_raises_error(self):
-        """
-        [シナリオ] なぞなぞモードで性別を指定せずに get_chat_history を呼び出す
-        [期待値] ValueError が発生すること
-        """
-        user_message = MessageDTO(
-            user=self.user,
-            role=RoleType.USER,
-            content="なぞなぞスタート",
-            model_name=ModelName.GPT_5_MINI,
-            use_case_type=UseCaseType.RIDDLE,
-        )
-        with self.assertRaisesRegex(ValueError, "gender is required for RiddleUseCase"):
-            ChatService.get_chat_history(
-                user_message, use_case_type=UseCaseType.RIDDLE, gender=None
-            )
-
-    def test_riddle_use_case_no_gender_raises_error(self):
-        """
-        [シナリオ] RiddleUseCase.execute を性別指定なしで呼び出す
-        [期待値] ValueError が発生すること
-        """
-        config = OpenAIGptConfig(
-            api_key="fake", max_tokens=100, model=ModelName.GPT_5_MINI
-        )
-        use_case = RiddleUseCase(config)
-        with self.assertRaisesRegex(ValueError, "gender is required for RiddleUseCase"):
-            use_case.execute(user=self.user, content="なぞなぞスタート", gender=None)
-
-
-class OpenAiUseCaseTest(TestCase):
-    def setUp(self):
-        self.user = User.objects.create_user(
-            username="openai_user", password="password"
-        )
-
-    def _assert_chat_log_saved(self, model_name: ModelName, use_case_type: UseCaseType):
-        """ChatLogs にファイルパスとモデル名、ユースケースタイプが正しく保存されていることを検証する共通ヘルパー"""
-        last_log = ChatLogs.objects.filter(
-            user=self.user, role=RoleType.ASSISTANT.value
-        ).last()
-        self.assertIsNotNone(last_log)
-        self.assertIsNotNone(last_log.file.name)
-        self.assertEqual(last_log.model_name, model_name)
-        self.assertEqual(last_log.use_case_type, use_case_type)
-
-    @patch("llm_chat.domain.service.completion.multimedia.OpenAILlmDalleService")
-    @patch("llm_chat.domain.service.completion.multimedia.requests.get")
-    @patch("llm_chat.domain.service.completion.multimedia.Image.open")
-    def test_dalle_use_case_saves_file_path(
-        self, mock_image_open, mock_get, mock_dalle_service
-    ):
-        """
-        [シナリオ: DALL-E 3画像生成]
-        1. OpenAIDalleUseCase を実行して画像を生成
-        2. 期待値:
-           - 返された MessageDTO に file_path が含まれていること
-           - DB (ChatLogs) にファイルパスとモデル名が正しく保存されていること
-        """
-        # DALL-E サービスと画像処理をモック化
-        mock_response = MagicMock()
-        mock_response.data = [MagicMock(url="http://example.com/image.jpg")]
-        mock_dalle_service.return_value.retrieve_answer.return_value = mock_response
-
-        mock_get_response = MagicMock()
-        mock_get_response.content = b"fake_image_content"
-        mock_get.return_value = mock_get_response
-
-        mock_img = MagicMock()
-        mock_image_open.return_value.resize.return_value = mock_img
-
-        # UseCase 実行
-        use_case = OpenAIDalleUseCase()
-        result = use_case.execute(self.user, "ねこの画像を生成して")
-
-        # 結果の MessageDTO を検証
-        self.assertIsNotNone(result.file_path)
-        self.assertEqual(result.model_name, ModelName.DALLE_3)
-
-        # DB への保存を検証
-        self._assert_chat_log_saved(ModelName.DALLE_3, UseCaseType.OPENAI_DALLE)
-
-    @patch("llm_chat.domain.service.completion.multimedia.OpenAILlmTextToSpeech")
-    def test_tts_use_case_saves_file_path(self, mock_tts_service):
-        """
-        [シナリオ: TTS音声生成]
-        1. OpenAITextToSpeechUseCase を実行して音声を生成
-        2. 期待値:
-           - 返された MessageDTO に file_path が含まれていること
-           - DB (ChatLogs) にファイルパスとモデル名が正しく保存されていること
-        """
-        # TTS サービスをモック化
-        mock_response = MagicMock()
-        mock_tts_service.return_value.retrieve_answer.return_value = mock_response
-
-        # UseCase 実行
-        use_case = OpenAITextToSpeechUseCase()
-        result = use_case.execute(self.user, "こんにちは")
-
-        # 結果の MessageDTO を検証
-        self.assertIsNotNone(result.file_path)
-        self.assertEqual(result.model_name, ModelName.TTS_1)
-
-        # DB への保存を検証
-        self._assert_chat_log_saved(ModelName.TTS_1, UseCaseType.OPENAI_TEXT_TO_SPEECH)
-
-    @patch("llm_chat.domain.service.completion.multimedia.OpenAILlmSpeechToText")
-    @patch("llm_chat.domain.service.completion.multimedia.Path.exists")
-    def test_stt_use_case_saves_file_path(self, mock_exists, mock_stt_service):
-        """
-        [シナリオ: STT音声認識]
-        1. OpenAISpeechToTextUseCase を実行して音声をテキスト化
-        2. 期待値:
-           - DB (ChatLogs) にファイルパスとモデル名が正しく保存されていること
-        """
-        # 前準備
-        mock_exists.return_value = True
-        mock_stt_service.return_value.retrieve_answer.return_value = MagicMock(
-            text="テスト音声です"
-        )
-
-        # ダミーのアップロードファイルを作成
-
-        audio_file = SimpleUploadedFile(
-            "test.mp3", b"dummy content", content_type="audio/mpeg"
-        )
-
-        # UseCase 実行
-        with patch("llm_chat.domain.use_case.completion.multimedia.open", create=True):
-            use_case = OpenAISpeechToTextUseCase(audio_file)
-            result = use_case.execute(self.user, "N/A")
-
-        # 結果の MessageDTO を検証
-        self.assertEqual(result.file_path, "llm_chat/audios/test.mp3")
-        self.assertEqual(result.model_name, ModelName.WHISPER_1)
-
-        # DB への保存を検証
-        last_log = ChatLogs.objects.filter(
-            user=self.user, role=RoleType.ASSISTANT.value
-        ).last()
-        self.assertIsNotNone(last_log)
-        self.assertEqual(last_log.file.name, "llm_chat/audios/test.mp3")
-        self.assertEqual(last_log.model_name, ModelName.WHISPER_1)
-        self.assertEqual(last_log.use_case_type, UseCaseType.OPENAI_SPEECH_TO_TEXT)
-
-
-class ViewLogicTest(TestCase):
-    def setUp(self):
-        self.user = User.objects.create_superuser(
-            username="admin", password="password", email="admin@example.com"
-        )
-
-    def test_index_view_riddle_active_status(self):
-        """
-        [シナリオ: Viewにおけるなぞなぞ活性状態判定]
-        1. 履歴がない場合: is_riddle_active が False であることを確認
-        2. 最新履歴がなぞなぞメッセージ (終了なし) の場合: is_riddle_active が True になることを確認
-        3. 最新履歴になぞなぞ終了メッセージが含まれる場合: is_riddle_active が False に戻ることを確認
-        """
-
-        factory = RequestFactory()
-        request = factory.get("/")
-        request.user = self.user
-        request.session = SessionStore()
-
-        # 1. 履歴なし -> Riddle 非アクティブ
-        view = IndexView()
-        view.request = request
-        context = view.get_context_data()
-        self.assertFalse(context["is_riddle_active"])
-
-        # 2. なぞなぞ開始メッセージあり -> Riddle アクティブ
-        ChatLogs.objects.create(
-            user=self.user,
-            role=RoleType.ASSISTANT.value,
-            content="なぞなぞを出題します",
-            use_case_type=UseCaseType.RIDDLE,
-        )
-        context = view.get_context_data()
-        self.assertTrue(context["is_riddle_active"])
-
-        # 3. なぞなぞ終了メッセージあり -> Riddle 非アクティブ
-        ChatLogs.objects.create(
-            user=self.user,
-            role=RoleType.ASSISTANT.value,
-            content=f"お疲れ様でした。 {RiddleChatService.RIDDLE_END_MESSAGE}",
-            use_case_type=UseCaseType.RIDDLE,
-        )
-        context = view.get_context_data()
-        self.assertFalse(context["is_riddle_active"])
-
-    def test_index_view_initial_use_case_type(self):
-        """
-        [シナリオ: IndexView の get_initial() による直近ユースケースタイプの取得]
-        1. 履歴がない場合: デフォルト値が返ることを確認
-        2. 直近の履歴が Gemini の場合: Gemini が返ることを確認
-        3. 直近の履歴が Dall-e の場合: OpenAIDalle が返ることを確認
-        4. 直近の履歴が なぞなぞ (進行中) の場合: Riddle が返ることを確認
-        5. 直近の履歴が なぞなぞ (終了) の場合: 直近の model_name に基づく値が返ることを確認
-        """
-
-        factory = RequestFactory()
-        request = factory.get("/")
-        request.user = self.user
-        request.session = SessionStore()
-
-        view = IndexView()
-        view.request = request
-
-        # 1. 履歴なし -> デフォルトで UseCaseType.OPENAI_GPT
-        initial = view.get_initial()
-        self.assertEqual(initial.get("use_case_type"), UseCaseType.OPENAI_GPT)
-
-        # 2. 直近が Gemini
-        ChatLogs.objects.create(
-            user=self.user,
-            role=RoleType.USER.value,
-            content="Hello",
-            model_name=ModelName.GEMINI_2_0_FLASH,
-            use_case_type=UseCaseType.GEMINI,
-            created_at=timezone.now(),
-        )
-        time.sleep(0.01)
-        initial = view.get_initial()
-        self.assertEqual(initial.get("use_case_type"), UseCaseType.GEMINI)
-
-        # 3. 直近が Dall-e
-        ChatLogs.objects.create(
-            user=self.user,
-            role=RoleType.ASSISTANT.value,
-            content="Image URL",
-            model_name=ModelName.DALLE_3,
-            use_case_type=UseCaseType.OPENAI_DALLE,
-            created_at=timezone.now(),
-        )
-        time.sleep(0.01)
-        initial = view.get_initial()
-        self.assertEqual(initial.get("use_case_type"), UseCaseType.OPENAI_DALLE)
-
-        # 4. なぞなぞ (進行中)
-        ChatLogs.objects.create(
-            user=self.user,
-            role=RoleType.ASSISTANT.value,
-            content="なぞなぞです",
-            model_name=ModelName.GPT_5_MINI,
-            use_case_type=UseCaseType.RIDDLE,
-            created_at=timezone.now(),
-        )
-        time.sleep(0.01)
-        initial = view.get_initial()
-        self.assertEqual(initial.get("use_case_type"), UseCaseType.RIDDLE)
-
-        # 5. なぞなぞ (終了) -> 最新ログが Riddle 終了であれば、Riddle 活性は False となり、
-        # 最新ログ (Riddle 終了メッセージ) の use_case_type は UseCaseType.RIDDLE だが、
-        # 以前の仕様では model_name に基づき OpenAIGpt が選択されていたが、
-        # 新しい仕様では終了後も最後の use_case_type である Riddle が選択される
-        # （なぞなぞが終了していても、最後になぞなぞをしていたなら次もなぞなぞモードで開始するのが自然なため）
-        ChatLogs.objects.create(
-            user=self.user,
-            role=RoleType.ASSISTANT.value,
-            content=f"正解です！ {RiddleChatService.RIDDLE_END_MESSAGE}",
-            model_name=ModelName.GPT_5_MINI,
-            use_case_type=UseCaseType.RIDDLE,
-            created_at=timezone.now(),
-        )
-        initial = view.get_initial()
-        self.assertEqual(initial.get("use_case_type"), UseCaseType.RIDDLE)
-
-        # 6. なぞなぞ進行中に別のチャットを挟む -> 最新のログがなぞなぞ以外であれば、その時点でなぞなぞは中断されているとみなす
-        ChatLogs.objects.create(
-            user=self.user,
-            role=RoleType.ASSISTANT.value,
-            content="なぞなぞ再開",
-            model_name=ModelName.GPT_5_MINI,
-            use_case_type=UseCaseType.RIDDLE,
-            created_at=timezone.now(),
-        )
-        time.sleep(0.01)
-        ChatLogs.objects.create(
-            user=self.user,
-            role=RoleType.USER.value,
-            content="横槍チャット",
-            model_name=ModelName.GEMINI_2_0_FLASH,
-            use_case_type=UseCaseType.GEMINI,
-            created_at=timezone.now(),
-        )
-        initial = view.get_initial()
-        # 最新が Gemini なので、Gemini が選択される
-        self.assertEqual(initial.get("use_case_type"), UseCaseType.GEMINI)
-
-        # 明示的に、なぞなぞを終了させる
-        ChatLogs.objects.create(
-            user=self.user,
-            role=RoleType.ASSISTANT.value,
-            content=f"お疲れ様でした。 {RiddleChatService.RIDDLE_END_MESSAGE}",
-            model_name=ModelName.GPT_5_MINI,
-            use_case_type=UseCaseType.RIDDLE,
-            created_at=timezone.now(),
-        )
-
-        # 7. ストリーミングモードの復元
-        ChatLogs.objects.create(
-            user=self.user,
-            role=RoleType.USER.value,
-            content="Streaming request",
-            model_name=ModelName.GPT_5_MINI,
-            use_case_type=UseCaseType.OPENAI_GPT_STREAMING,
-            created_at=timezone.now(),
-        )
-        initial = view.get_initial()
-        self.assertEqual(initial.get("use_case_type"), UseCaseType.OPENAI_GPT_STREAMING)
-
-        # 8. RAGモードの復元
-        ChatLogs.objects.create(
-            user=self.user,
-            role=RoleType.USER.value,
-            content="RAG query",
-            model_name=ModelName.GPT_5_MINI,
-            use_case_type=UseCaseType.OPENAI_RAG,
-            created_at=timezone.now(),
-        )
-        initial = view.get_initial()
-        self.assertEqual(initial.get("use_case_type"), UseCaseType.OPENAI_RAG)
-
-        # 9. セッションからの use_case_type 復元 (DB 履歴より優先)
-        request.session["use_case_type"] = UseCaseType.GEMINI
-        initial = view.get_initial()
-        self.assertEqual(initial.get("use_case_type"), UseCaseType.GEMINI)
-
-        # 10. なぞなぞ進行中が最優先 (セッションより優先)
-        ChatLogs.objects.create(
-            user=self.user,
-            role=RoleType.ASSISTANT.value,
-            content="なぞなぞです",
-            model_name=ModelName.GPT_5_MINI,
-            use_case_type=UseCaseType.RIDDLE,
-            created_at=timezone.now() + timedelta(seconds=1),
-        )
-        initial = view.get_initial()
-        self.assertEqual(initial.get("use_case_type"), UseCaseType.RIDDLE)
-
-    def test_sync_response_view_riddle_no_questions_error(self):
-        """
-        [シナリオ] なぞなぞの問題が登録されていない状態で SyncResponseView にリクエストを投げる
-        [期待値] 400エラーが返り、適切なメッセージが含まれていること
-        """
-        RiddleQuestion.objects.all().delete()
-        factory = RequestFactory()
-        request = factory.post(
-            "/llm_chat/sync/",
-            {
-                "use_case_type": UseCaseType.RIDDLE,
-                "user_input": "スタート",
-                "gender": "man",
-            },
-        )
-        request.user = self.user
-        # セッションのモック
-        request.session = SessionStore()
-
-        from llm_chat.views import SyncResponseView
-
-        response = SyncResponseView.post(request)
-        self.assertEqual(response.status_code, 400)
-
-        import json
-
-        data = json.loads(response.content)
-        self.assertEqual(
-            data["error"],
-            "なぞなぞの問題が登録されていません。管理画面から問題を登録してください。",
-        )
-
-    def test_riddle_sample_csv_view(self):
-        """
-        [シナリオ] サンプルCSVダウンロードリンクにアクセスする
-        [期待値] 200 OK が返り、CSV形式で期待通りの内容が含まれていること
-        """
-        from django.urls import reverse
-        from llm_chat.views import RiddleSampleCSVView
-
-        factory = RequestFactory()
-        request = factory.get(reverse("llm:riddle_sample_csv"))
-        request.user = self.user
-
-        response = RiddleSampleCSVView.get(request)
-
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response["Content-Type"], "text/csv")
-        self.assertIn(
-            'attachment; filename="riddle_sample.csv"', response["Content-Disposition"]
-        )
-
-        content = response.content.decode("utf-8")
-        lines = content.strip().splitlines()
-        self.assertEqual(len(lines), 2)
-        self.assertIn("はじめは4本足", lines[0])
-        self.assertIn("人間", lines[0])
-        self.assertIn("私は黒い服を着て", lines[1])
-        self.assertIn("たいまつ", lines[1])
-
-    def test_riddle_csv_upload_overwrite(self):
-        """
-        [シナリオ] CSVをアップロードして、既存の問題を全削除して再投入する。
-        [期待値]
-        - 既存のデータは消え、CSVの内容が登録される。
-        - order や question_text が重複していても、トランザクションが効いて全削除されるため正常。
-        """
-        from django.urls import reverse
-        from llm_chat.views import RiddleCSVUploadView
-        import io
-
-        # 1. 初期データ作成
-        RiddleQuestion.objects.all().delete()
-        RiddleQuestion.objects.create(
-            order=1,
-            question_text="古い問題",
-            answer_text="答え1",
-        )
-
-        # 2. CSVデータ準備 (既存とは全く異なるデータ 1件 + 既存を上書きする内容 1件)
-        # order, question_text, answer_text
-        csv_content = [
-            "1,上書き問題1,新しい答え",
-            "2,新規問題2,答え2",
-        ]
-        csv_file = io.BytesIO("\n".join(csv_content).encode("utf-8"))
-        csv_file.name = "test.csv"
-
-        factory = RequestFactory()
-        request = factory.post(
-            reverse("llm:riddle_csv_upload"),
-            {"csv_file": csv_file},
-        )
-        request.user = self.user
-        request.session = SessionStore()
-        request._messages = self._create_messages_mock()
-
-        # 3. 実行
-        from llm_chat.views import RiddleCSVUploadView
-
-        response = RiddleCSVUploadView.post(request)
-
-        # 4. 検証
-        self.assertEqual(response.status_code, 302)  # リダイレクト
-        self.assertEqual(RiddleQuestion.objects.count(), 2)
-
-        # 既存データ「古い問題」が消えていることを確認（全上書き方式のため）
-        self.assertFalse(
-            RiddleQuestion.objects.filter(question_text="古い問題").exists()
-        )
-
-        # CSVの内容が登録されていることを確認
-        q1 = RiddleQuestion.objects.get(order=1)
-        self.assertEqual(q1.question_text, "上書き問題1")
-        self.assertEqual(q1.answer_text, "新しい答え")
-
-        q2 = RiddleQuestion.objects.get(order=2)
-        self.assertEqual(q2.question_text, "新規問題2")
-
-    def test_riddle_csv_upload_complex_reorder(self):
-        """
-        [シナリオ]
-        1. order 1, 2, 3 で登録する。
-        2. 既存の 2 を消し、新たに 2 と 4 を登録する（1, 2, 3 -> 1, 3, 2, 4 のような順序変更を伴うケース）。
-        [期待値]
-        - 全削除・再登録方式により、制約違反を起こさずに期待通りの順序で登録される。
-        """
-        from django.urls import reverse
-        from llm_chat.views import RiddleCSVUploadView
-        import io
-
-        # 1. 初期データ作成 (1, 2, 3)
-        RiddleQuestion.objects.all().delete()
-        RiddleQuestion.objects.create(order=1, question_text="Q1", answer_text="A1")
-        RiddleQuestion.objects.create(order=2, question_text="Q2", answer_text="A2")
-        RiddleQuestion.objects.create(order=3, question_text="Q3", answer_text="A3")
-
-        # 2. CSVデータ準備 (1, 3, 2, 4)
-        # ユーザーの意図: "2を消して、2と4を新しく入れる"
-        # これは CSV 上では 1, 3, 2(新), 4(新) という並びや、単純に 1, 2(新), 3, 4(新) という並びになる。
-        # ここでは順序の入れ替えが発生するケースをテストする。
-        csv_content = [
-            "1,Q1,A1",
-            "3,Q3,A3",
-            "2,New Q2,New A2",
-            "4,New Q4,New A4",
-        ]
-        csv_file = io.BytesIO("\n".join(csv_content).encode("utf-8"))
-        csv_file.name = "test_complex.csv"
-
-        factory = RequestFactory()
-        request = factory.post(
-            reverse("llm:riddle_csv_upload"),
-            {"csv_file": csv_file},
-        )
-        request.user = self.user
-        request.session = SessionStore()
-        request._messages = self._create_messages_mock()
-
-        # 3. 実行
-        from llm_chat.views import RiddleCSVUploadView
-
-        response = RiddleCSVUploadView.post(request)
-
-        # 4. 検証
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(RiddleQuestion.objects.count(), 4)
-
-        # 順序通りに並んでいるか確認
-        questions = RiddleQuestion.objects.all().order_by("order")
-        self.assertEqual(questions[0].order, 1)
-        self.assertEqual(questions[0].question_text, "Q1")
-
-        self.assertEqual(questions[1].order, 2)
-        self.assertEqual(questions[1].question_text, "New Q2")
-
-        self.assertEqual(questions[2].order, 3)
-        self.assertEqual(questions[2].question_text, "Q3")
-
-        self.assertEqual(questions[3].order, 4)
-        self.assertEqual(questions[3].question_text, "New Q4")
-
-        # Q2 が更新されている（古い Q2 は消えている）ことを確認
-        self.assertFalse(RiddleQuestion.objects.filter(question_text="Q2").exists())
-
-    @staticmethod
-    def _create_messages_mock():
-        from django.contrib.messages.storage.base import BaseStorage
-        from django.test import RequestFactory
-
-        request = RequestFactory().get("/")
-        return BaseStorage(request)
