+++ llm_chat/tests/domain/factory/completion/test_use_case.py
